Related work items: #12947, #12948, #12950, #12951, #13098, #13099, #13100, #13101, #13104, #13106, #13107, #13108, #13109, #14889

+++ dotnet/corecomponents/net.atos.daf.ct2.reports/entity/UserPreferenceReportDataColumn.cs
@@ -23,7 +23,7 @@
         public string ChartType { get; set; }
         public string ReportPreferenceType { get; set; }
         public string ThresholdType { get; set; }
-        public long ThresholdValue { get; set; }
+        public double ThresholdValue { get; set; }
         public int[] SubDataAttributes { get; set; }
         private char AttributeType { get; set; }
         public int? TargetProfileId { get; set; }
